requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
frozenlist==1.4.1
h11==0.14.0
idna==3.10
multidict==6.1.0
pillow==10.4.0
propcache==0.2.0
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
wsproto==1.2.0
yarl==1.14.0
